refactor(checker): route type parameter default relation guards
